+++ test/test_helpers.dart
@@ -13,7 +13,7 @@ void createDirs(List<Directory> dirs) {
 }
 
 Directory createTempDir(String suffix, [String? folderName]) {
-  Directory dir = Directory.systemTemp.createTempSync(suffix);
+  final dir = Directory.systemTemp.createTempSync(suffix);
 
   if (!dir.existsSync()) {
     dir.createSync(recursive: true);
@@ -24,7 +24,7 @@ Directory createTempDir(String suffix, [String? folderName]) {
     return dir;
   }
 
-  Directory newDir = Directory(join(dir.path, folderName));
+  final newDir = Directory(join(dir.path, folderName));
   if (!newDir.existsSync()) {
     newDir.createSync(recursive: true);
   }
@@ -39,5 +39,35 @@ void deleteDirs(List<Directory> dirs) {
       dir.deleteSync(recursive: true);
     }
     expect(dir.existsSync(), isFalse);
+  }
+}
+
+/// Recursively copy the contents of [source] into [destination].
+///
+/// Existing files in [destination] will be overwritten.
+void copyDirectory(Directory source, Directory destination) {
+  if (!source.existsSync()) {
+    throw ArgumentError('Source directory does not exist: ${source.path}');
+  }
+
+  if (!destination.existsSync()) {
+    destination.createSync(recursive: true);
+  }
+
+  for (final entity in source.listSync(recursive: true, followLinks: false)) {
+    final relativePath = relative(entity.path, from: source.path);
+    final newPath = join(destination.path, relativePath);
+
+    if (entity is Directory) {
+      final newDir = Directory(newPath);
+      if (!newDir.existsSync()) {
+        newDir.createSync(recursive: true);
+      }
+    } else if (entity is File) {
+      final newFile = File(newPath);
+      newFile
+        ..createSync(recursive: true)
+        ..writeAsBytesSync(entity.readAsBytesSync());
+    }
   }
 }